edf: correct capture against AS11 clock drift
- measure the AS11 offset against NTP with monotonic RPC timing
- freeze one transform for headers, samples, events, and STR records
- keep therapy, mask, and _ZLE boundaries as separate events
- preserve the native local EDF axis after correcting absolute time

+++ include/runtime_snapshots.h
@@ -37,7 +37,7 @@ struct DeviceSnapshot {
     bool ready = false;
     As11TherapyState therapy_state = As11TherapyState::Unknown;
     bool clock_valid = false;
-    int32_t clock_offset_ms = 0;
+    int64_t clock_offset_ms = 0;
     int32_t timezone_offset_minutes = 0;
     uint32_t generation = 0;
 };
